Fix edge case in indentation checks
If getFirstNonEmptyLineno() returns the value of the last line
number, this doesn't guarantee that the last line is in fact empty
since the firstNonEmptyLineno variable was set to that value at the
beginning of the function. Give firstNonEmptyLineno an unambiguous
starting value of -1 to know for sure if the for loop doesn't
actually encounter any non-empty lines.